feat(agentic): add mode-scoped subagent visibility
- introduce builtin subagent visibility policies to restrict subagent exposure by parent mode
- keep review and research subagents visible only to their intended built-in modes
- reorganize agent definitions and split registry modules to support clearer visibility and registration boundaries
- update desktop and web subagent integration to reflect the new visibility-aware registry behavior
