--- src/crates/core/src/agentic/agents/custom_subagents/mod.rs
@@ -1,5 +1,0 @@
-mod custom_subagent;
-mod custom_subagent_loader;
-
-pub use custom_subagent::{CustomSubagent, CustomSubagentKind};
-pub use custom_subagent_loader::CustomSubagentLoader;
