+++ src/crates/core/src/agentic/agents/definitions/custom/mod.rs
@@ -1,0 +1,3 @@
+mod subagent;
+
+pub use subagent::{CustomSubagent, CustomSubagentKind};
