+++ src/crates/core/src/agentic/agents/definitions/hidden/mod.rs
@@ -1,0 +1,9 @@
+mod code_review;
+mod deep_review;
+mod generate_doc;
+mod init;
+
+pub use code_review::CodeReviewAgent;
+pub use deep_review::DeepReviewAgent;
+pub use generate_doc::GenerateDocAgent;
+pub use init::InitAgent;
