+++ src/crates/core/src/agentic/agents/definitions/mod.rs
@@ -1,0 +1,6 @@
+pub mod custom;
+pub mod hidden;
+pub mod modes;
+pub mod review;
+pub mod shared;
+pub mod subagents;
