+++ src/crates/core/src/agentic/agents/definitions/modes/claw.rs
@@ -1,18 +1,18 @@
-//! Agentic Mode
+//! Claw Mode
 
-use super::Agent;
+use crate::agentic::agents::{Agent, RequestContextPolicy};
 use async_trait::async_trait;
-pub struct AgenticMode {
+pub struct ClawMode {
     default_tools: Vec<String>,
 }
 
-impl Default for AgenticMode {
+impl Default for ClawMode {
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl AgenticMode {
+impl ClawMode {
     pub fn new() -> Self {
         Self {
             default_tools: vec![
@@ -25,12 +25,16 @@ impl AgenticMode {
                 "Grep".to_string(),
                 "Glob".to_string(),
                 "WebSearch".to_string(),
-                "TodoWrite".to_string(),
-                "GenerativeUI".to_string(),
                 "Skill".to_string(),
-                "AskUserQuestion".to_string(),
                 "Git".to_string(),
                 "TerminalControl".to_string(),
+                "SessionControl".to_string(),
+                "SessionMessage".to_string(),
+                "SessionHistory".to_string(),
+                "Cron".to_string(),
+                // Browser, terminal, and routing metadata live under ControlHub.
+                // Local desktop/system control is delegated to the ComputerUse
+                // agent/tool instead of being surfaced as a ControlHub domain.
                 "ControlHub".to_string(),
             ],
         }
@@ -38,52 +42,36 @@ impl AgenticMode {
 }
 
 #[async_trait]
-impl Agent for AgenticMode {
+impl Agent for ClawMode {
     fn as_any(&self) -> &dyn std::any::Any {
         self
     }
 
     fn id(&self) -> &str {
-        "agentic"
+        "Claw"
     }
 
     fn name(&self) -> &str {
-        "Agentic"
+        "Claw"
     }
 
     fn description(&self) -> &str {
-        "Full-featured AI assistant with access to all tools for comprehensive software development tasks"
+        "Personal assistant for daily tasks"
     }
 
     fn prompt_template_name(&self, _model_name: Option<&str>) -> &str {
-        "agentic_mode"
+        "claw_mode"
     }
 
     fn default_tools(&self) -> Vec<String> {
         self.default_tools.clone()
     }
 
+    fn request_context_policy(&self) -> RequestContextPolicy {
+        RequestContextPolicy::full_without_layout()
+    }
+
     fn is_readonly(&self) -> bool {
         false
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{Agent, AgenticMode};
-
-    #[test]
-    fn always_uses_default_prompt_template() {
-        let agent = AgenticMode::new();
-        assert_eq!(agent.prompt_template_name(Some("gpt-5.1")), "agentic_mode");
-        assert_eq!(
-            agent.prompt_template_name(Some("GPT-5-CODEX")),
-            "agentic_mode"
-        );
-        assert_eq!(
-            agent.prompt_template_name(Some("claude-sonnet-4")),
-            "agentic_mode"
-        );
-        assert_eq!(agent.prompt_template_name(None), "agentic_mode");
     }
 }
