+++ src/crates/core/src/agentic/agents/definitions/modes/mod.rs
@@ -1,0 +1,15 @@
+mod agentic;
+mod claw;
+mod cowork;
+mod deep_research;
+mod debug;
+mod plan;
+mod team;
+
+pub use agentic::AgenticMode;
+pub use claw::ClawMode;
+pub use cowork::CoworkMode;
+pub use deep_research::DeepResearchMode;
+pub use debug::DebugMode;
+pub use plan::PlanMode;
+pub use team::TeamMode;
