+++ src/crates/core/src/agentic/agents/definitions/review/mod.rs
@@ -1,0 +1,8 @@
+mod review_fixer;
+mod review_specialists;
+
+pub use review_fixer::ReviewFixerAgent;
+pub use review_specialists::{
+    ArchitectureReviewerAgent, BusinessLogicReviewerAgent, FrontendReviewerAgent,
+    PerformanceReviewerAgent, ReviewJudgeAgent, SecurityReviewerAgent,
+};
