+++ src/crates/core/src/agentic/agents/definitions/shared/mod.rs
@@ -1,0 +1,3 @@
+mod readonly;
+
+pub use readonly::ReadonlySubagent;
