+++ src/crates/core/src/agentic/agents/definitions/subagents/explore.rs
