+++ src/crates/core/src/agentic/agents/definitions/subagents/mod.rs
@@ -1,0 +1,9 @@
+mod computer_use;
+mod explore;
+mod file_finder;
+mod research_specialist;
+
+pub use computer_use::ComputerUseMode;
+pub use explore::ExploreAgent;
+pub use file_finder::FileFinderAgent;
+pub use research_specialist::ResearchSpecialistAgent;
