+++ src/crates/core/src/agentic/agents/definitions/subagents/research_specialist.rs
